RCMS/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-SocketIO
paramiko
PyYAML
psutil
